backend/requirements.txt
ultralytics==8.3.23
torch>=2.1.0
torchvision>=0.16.0
opencv-python==4.10.0.84
numpy>=1.23.0
easyocr==1.7.1
Pillow>=10.0.0

Flask==3.0.3
Flask-Cors==5.0.0

pymongo==4.10.1
dnspython>=2.6.0
python-dotenv==1.0.1

gunicorn==23.0.0


# Auth & security
PyJWT==2.10.1
bcrypt==4.2.1

